Extract setAmendmentFields logic into a function and add unit tests
